requirements.txt
asgiref==3.2.10
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
clickclick==1.2.2
connexion==2.7.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.10
django-filter==2.3.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-jsonapi==3.2.0
djoser==2.0.5
Flask==1.1.2
idna==2.10
inflection==0.5.1
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
openapi-spec-validator==0.2.9
Pygments==2.7.1
pyrsistent==0.17.3
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==1.0.1
